+++ src/Foundation/Support/vanilo2_compatibility.php
@@ -1,5 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-\class_alias(BaseException::class, \DaveJamesMiller\Breadcrumbs\BreadcrumbsException::class);