+++ Changelog.md
@@ -10,6 +10,7 @@
 - Removed Admin from "Framework" - it is available as an optional separate package see [vanilo/admin](https://github.com/vanilophp/admin)
 - Minimum Laravel version is 8.22.1. [See GHSA-3p32-j457-pg5x](https://github.com/advisories/GHSA-3p32-j457-pg5x)
 - BC: Renamed Framework Module and Namespace to "Foundation"
+- BC: renamed `vanilo.framework.*` config values to `vanilo.foundation.*`
 
 ## 2.x Series